services

+++ models/ProdutoModel.cs
@@ -3,7 +3,7 @@ namespace firstORM.models
     public class ProdutoModel{
         public int id {get;set;}
         public String nome {get;set;}
-        public double valor {get;set;}
+        public decimal valor {get;set;}
 
         public String fornecedor {get; set;}
     }
